Refactored `Module` to be part of the Library Generation
The `Module` struct is used in no other crate, so there is no need for
the entire `rusty-junctions-utils` crate.
We can include the `Module` in the proc-macro Library Generation
crate.
